Extract InLineupScore class
